Rejestracja w jsp z walidacją danych i enumem.

+++ src/main/java/pl/mygames/hackandslash/dao/IGenericExtendedDao.java
@@ -37,7 +37,7 @@ public interface IGenericExtendedDao<T, PK extends Serializable> {
 
     T findById(PK id) throws EmptyResultDataAccessException;
 
-    Integer count() throws EmptyResultDataAccessException; 
+    Long count() throws EmptyResultDataAccessException; 
     
     Integer generateId();
